added compassCalibration, for calibrating live

+++ src/calculates.h
@@ -1,11 +1,11 @@
 class calculate {
     public:
-    int rotationTime(int diam, int distance, float rot_speed);
+    long rotationTime(int diam, int distance, float rot_speed);
 };
 
-int calculate::rotationTime (int diam, int distance, float rot_speed){
+long calculate::rotationTime (int diam, int distance, float rot_speed){
     float rot_numb = distance / (diam * 3.14);
-    int time = (rot_numb / rot_speed) * 1000;
+    long time = (rot_numb / rot_speed) * 1000;
     return time;
 }
